requirements.txt
# 依赖文件

Appium-Python-Client==3.2.0
pytest==8.3.4
pytest-xdist==3.6.1
allure-pytest==2.13.5
PyYAML==6.0.2

# 更新所有依赖
# pip install -r requirements.txt --upgrade
